Move start node operations to api.
